feat: variables
feat: subtraction builtin

+++ src/interpreter/objects/mod.rs
@@ -1,7 +1,7 @@
 #[derive(Debug, Clone)]
 pub enum Object {
     List(Vec<Object>),
-    Integer(usize),
+    Integer(isize),
     String(String),
     Bool(bool),
     Builtin(fn(Vec<Object>) -> Object),
